#7107: move new test into existing test class
